Move the hack module to its own file
